Changed Client::shutdown() so it doesn't hold open queuesLock across the whole of shutdown. Otherwise we can have a two-mutex deadlock between queuesLock and handling_mutex during shutdown. See code comment for more details.
Move my_basename_r into pist_syslog.h so it can be used by other code
beyond eventmeth.cc.

Moved GUARD_AND_DBG_LOG into pist_check.h, so it can be used by other
code beyond eventmeth.cc.
